Corrected defects in JSweet implementation of core
First, the JS-implemented fields were not implementing toString() with
no arguments correctly.  This caused spurious differences when checking
side-effects during debug.

Second, the default j4ts implementation of Arrays.sort() does not dispatch
correctly when overloaded, in the case of an array of strings.  To force
the issue, I simply avoided the overloading by using the more general
form with an explicit comparator.  This works in Java as well, of course.