True full-screen mode toggle
This does not work on iPhones, where the Fullscreen API is unsupported as
of now.

Also, it does not correctly switch to the FullscreenExit icon.  To do that
correctly, it looks like I'll need to replicate this:

   https://github.com/Darth-Knoppix/example-react-fullscreen

Finally, it is a bit hacky with prop-drilling, since it wasn't working
inside the shadow dom.  There seem to be indications that it should work
OK with shadow dom, based on references to shadow root here:

   https://developer.mozilla.org/en-US/docs/Web/API/Fullscreen_API