Initial POC using WebXR via react-xr
The Poimandres team comes through again!  They have built in support for
WebXR:

  https://github.com/pmndrs/react-xr

I had to set up the dev server for HTTPS, in order for the Meta browser
to accept the server.  I used this page:

  https://nodejs.org/en/knowledge/HTTP/servers/how-to-create-a-HTTPS-server/

With just a few code changes (on the branch where I upgraded React and R3F),
I can now switch to VR mode in vZome Online (dev server).  The result is
a view floating in space at the origin, amidst vZome models of heroic
proportions (balls over 5 ft across).  Head tracking and movement both
work, though the lighting should not be coupled to the camera in VR.

There is a lot to do to reach UX parity with the native Unity code, but
it is pretty amazing that it works already with so few changes!