Preview button for VR mode
When viewing a web page with multiple vzome-viewer elements on the Oculus,
I want to be able to open any of the designs in VR, and currently only
one of the VRCanvas elements will work.

To solve this, I'm adding an "open in browser" button that launches vZome
Online in legacy preview mode.  It only shows if VR is available.

This also serves as a "preview" of the "open in vZome Online" that I
intend to provide for all viewers.