Added support for color changes
I also tried out loadUrl, but just as a temporary hack.  Since there
seems to be no access to the Android system clipboard, I'm not sure how
I could get URLs into the app.  That leaves me with no option except to
implement a GitHub browser, which I'm not going to do, or figure out
how to access the Downloads folder.