Updated to Unity 2021.3.6f1 LTS and XR Interaction Toolkit
These videos were most helpful:

  https://youtu.be/yxMzAw2Sg5w
  https://youtu.be/BZt74PVb7sM

I still have red/green block hand visualizations, and the movement is
disabled at the moment, but I can trigger the UI and the Java code runs
successfully.

There was a problem with scheduling the load file job, but I added an
explicit call to JobHandle .ScheduleBatchedJobs(), and that fixed it.