Enabled batch rendering for loading in VR
Rendering incrementally during load is way too expensive, as I knew even
for the desktop app... you watch while all the branches are explored,
only to be undone.  Therefore, I've implemented batch rendering for the
loading phase.

I still had to attach an incremental render listener after the load, so
that the undo/redo/undo-all buttons work correctly.  I intend to dispense
with those buttons, and instead implement a load/clear modality, so that
only one model is present at a time.  This is obviously for a viewer app,
not an editor.

I also disabled the "look behind" function on the right controller hat
switch.