Corrected side-effects and history exporters
Note: this may break some regression tests that compare history.

Manifestation.getXml() was exporting "NoConstructions" due to mutations
of the mManifests list later in the history.  A simple correction is to
always generate a Construction if the list is empty.

I need to do this in order to correctly compare the batch side-effects
export from Java with the incremental side-effects export in Javascript.
The incremental export happens before subsequent edits mutate the
mManifests lists, so it never generates "NoConstructions".