Correct threading for DocumentFrame finishLoading
When using the DapAdapter (with the VS Code debugger), we were getting a
race condition that could cause an NPE.  It was triggered because the
map of document controllers was getting modified on a background thread,
when the document was set to unnamed due to a migration, before the EDT
could capture the controller for later debugging actions.

I modified the DapAdapter to use the EDT for code that normally uses
it, to open a document.  Doing this enables serialization with other EDT
code, notably the document renaming when a migration happens.

For the same reason, at the end of the DocumentFrame constructor, I
changed the ExclusiveAction to a vanilla SwingWorker, since there is more
to do in the done() call than just releasing the excluder.

Now, the only thing that happens on the worker is the actual "finish.load"
action, and the popup dialogs, title change on migration, etc. are all
scheduled for the EDT.