Rationalize camera state
First, the camera state is now consistently stored in the Redux store...
no more local state.  This is essential.  Note that the *actual* camera
state, under the control of the TrackballControls, still gets out of sync,
so I still need to capture that state somehow whenever I'm going to
update the store, for example to toggle perspective.

Next, the camera state is now an exact copy of the vZome Camera state,
since that approach is eminently sensible for switching between
perspective and orthographic cameras.  The key is the "width" value,
which is the width of the view frustum at they eye distance.  When we
switch to perspective, using this width (which is in world coordinates)
keeps the view as stable as possible.

The camera position and field-of-view are now computed just in time.
Aspect ratio is also computed only when the measurements change.

Defects outstanding:

1. resizing the window does really wacky things to the ortho camera,
    with zoom out and a severe aspect ratio

2. toggling perspective resets the view orientation, since the live
    camera state is not captured