WIP Refactoring all exporters
This branch got started because I was trying to transpile DocumentModel
with JSweet.  I concluded that exporter classes should depend on DM, not
vice-versa.

The new approach looks good, but we really need Exporter3d exportScene(),
to be called only on exporters that are able to work with an article page.
This will require renaming of Exporter3d.needsManifestations() as
requiresFullDocument(), and use of it.  As the code stands, all exports
are the full document, never currentSnapshot and the current camera...
see the diff in DocumentController.doAction().