Tweaking the app name and version
Unfortunately, jpackage does not give me all the control I'd like.  In
order to make the DMG name correct, I have to skip the version in the
.app name, so it becomes just "vZome.app".

The menu bar name is still versioned, since that is a separate property.

I tried to set "appImage", but that is something else.