Fixed 73 JSweet transpile errors
JSweet seems incapable of dealing with the Iterable interface, probably
because of the default method implementation for forEach().  Therefore,
we had 73 errors saying things like "IcosahedralSymmetry is not assignable
to Symmetry".

Use of Iterable is OK; it is just implementation of it that is a problem.

The solution was to return an Iterable rather than implementing one.

This is yet another case of "un-clevering" the Java code, effectively
reducing our usable Java language to the subset that JSweet will accept.

More than half of the remaining 36 errors seem to be the same problem,
for different classes.