Aligned with newer transpile patterns
AbstractSymmetry manifested a true defect in JSweet, wherein the overloaded
constructor code reset the initialized fields at the end.  If you see
mDirectionMap get emptied after createFrameOrbits, this is the problem.

The other issue was a change in the pattern of interface testing, using
the __interfaces field, or possibly a change in the pattern of class
compilation, or both.  I had to make all my "manual" Java classes and
objects comply.