Reintroduced removed JRE classes
Since I'll have to split the jsweet build into two modules, one for the webworker and one for the main
context, I still need those JRE stubs that fill in gaps in `j4ts`.

This commit is to establish a working baseline for the webworker JSweet build, before I add a second
Gradle task for the UI context bundle.