WIP Multiple JSweet tasks
The JSweetPlugin for Gradle implements a single instance of JSweetPluginExtension that gets
added to the project.  This means that you can have multiple tasks of type JSweetTranspileTask,
but they all have identical configuration!

To resolve this, I'm no longer applying the plugin, and I've inlined its apply method in
build.gradle, with some necessary changes so names resolve and don't collide.  With that done,
I can attach different extension instances to the two tasks.