Data flow for Classic app and OrbitSetController
The data is flowing from the React component to the store, on to the worker, back to the
store, and finally back to the component state.

I have created some nice hooks for encapsulating the nested controllers: `useController`,
`useControllerList`, and `controllerAction`.  I'm not sure if all of this will be useful
in creating the WorkerStoreController that will be needed once I have legacy UI classes,
but we'll see.

Both property/list request responses and property change events come back to the store
(in the main browser context) as `CONTROLLER_PROPERTY_CHANGED` events, which are
turned into state changes by the reducer.

I have adopted the `event.meta==='WORKER'` approach to eliminate missing switch cases
in `workerSender`.