WIP Adding OrbitPanel to Classic app
After adding the missing `candy-metadata.json` files to `j4ts-file` and `j4ts-awt-swing`,
they now get included in the TSC phase correctly, and I get only 20 TS errors.  These are
all about legit gaps in `j4ts-awt-swing`, methods simply not implemented.

I've added two more Swing classes, bringing that down to 8 TS errors!

I'm now at the point of actually running OrbitPanel code!  It fails somewhere inside of
GridLayout.  I thought it might be because it does not have an HTML element bound, as if
it were in a JFrame or an Applet, so I tried `online/src/ui/java/vzome/OrbitPanelFrame.java`,
ala `j4ts-swing-example`, but just being a JFrame did not resolve the problem.

I've determined that it will be complicated to gain access to the Redux store from a JFrame
class, so I've switched to an applet strategy, creating and binding the applet from within
React.

I'm able to do `graphics.drawString()` on my panel, but stroking paths is not working
correctly, and there seem to be lots of gaps in the Graphics2D implementation.  I do, at least,
get part of my triangle drawn on the screen!