OrbitsApplet hooked up to worker
I've done a quick-and-dirty connection to the store, and it seems pretty awkward.

It is not working correctly yet, requiring a resize to trigger a repaint that reveals the dots.
However, I'm not going to invest any more time in `controller.js`, which has no future
without `j4ts-awt-swing` and the vZome desktop UI classes.  I've decided that this
approach is a non-starter for two reasons:

1. Renaud Pawlak confirms that `j4ts-awt-swing` is nowhere near production-ready
2. vZome desktop UI classes are not prepared to deal with completely async controllers,
  and `getProperty()` can't be made synchronous when it has to go to the worker and back.