PolytopesController decoupled from DocumentModel
For JSweet and the web, I don't want DocumentModel, at least for now,
since I have my own implementation on the JS side, and it is too
monolithic anyway.

In support of this, `Context` now includes `doEdit`, and a context argument
is passed to the `PolytopesController` constructor.

I removed a few methods from `DocumentModel`, to help reduce dependencies
on it.

`EditorModel` now extends `SymmetryAware`, and also has
`addSelectionSummaryListener`.

I'm now transpiling most of the AWT-free controllers.