Controller actions and property listeners
I'm setting up PropertyChangeListeners in a way that assumes all PCEs
are "doorbells", requiring a fetch of one or more properties.  This will
be a little inefficient for "legitimate" PCEs, which could be simply
forwarded, but should work fine.

I have also merged the property and commandList mechanisms, adding
an `isList` parameter.

The "predefinedOrbits" action is working correctly now.