WIP Using SolidJS at the root
Use of React is now confined to the DesignCanvas.  Only properties
flow from Solid to React, so I don't need to worry about context.

I've completely redesigned the worker, store, and controllers on the
client side.  Now each controller provides access to the store, and there
is no more global worker or state.

I still have only dealt with the classic app.  It seems to be working correctly,
though I have disabled parts of the nav bar for now.

All other apps and the web component are going to be broken at the moment.