Defect fix: SelectAll must explicitly have originLast set
See the two new regression tests.  After the change in behavior at commit
46f747cb43a99e99a70cff3959fd695f01abf5e2, the "false" variant was
failing (versions 7.1.3-7.1.5).  It is unusual for someone to use select-all
for an oriented purpose, but it can happen accidentally, as seen in
https://david-hall.github.io/vzome-sharing/2022/07/11/Online-rendering-and-side-effects-bug-01-10-32.html

Now, if someone has build a design that
depends on the implicit originLast behavior (a la the "true" variant), using
7.1.3-7.1.5, it will fail with 7.1.6 or later.  I'm prepared to make any such
fixes manually, adding the originLast attribute.