Fix: Non-maximized window size too big at startup
This fixes an issue that initially showed up when we switched desktop to Java 17. The main window is correctly maximized at startup, but when it's later switched to normal size (not maximized or minimized) then it was much bigger than the actual screen, at least on Windows 10. I'm not sure about the MAC. The solution is to set the "normal" size after the frame is made visible in the same way as commit 0ff6a4f41ac0d655c203930b5b0a9ba04d215267 did for the maximized window state.