Implemented click to select
This use-case is driving the full implementation of controller actions, including
with parameters.  I've updated `DefaultController` with `doParamAction()`
and `paramActionPerformed()` methods.  These enable the JS implementations
of `EditorController` and `PickingController`.

I eliminated the "i" prefix for instance IDs, as it provides no real value.  The prefix
for shape IDs is still useful to distinguish them from instance IDs.

Just for grins, the design automatically creates a 120-cell.