Integrated into downloads popup menu
I've tried replacing the CDN import for GLTFExporter with `three-stdlib`
but it does not give me the complete scene in the exported file.
I suspect there is a version disparity.

In any case, I'll leave it like this as it is working fine.

MUI styles are updated so the downloads button works in multiple
web components on a page.