Ability to switch planes
Everything seem to be correct!

Switching planes is working.  The hinge is preserved because it is
independent state.

There is a defect around camera dragging when the release
is interpreted as a plane click.

I have modified the way buildPlanes are constructed in the worker,
to record plane orientations.  Strut building now takes the plane
orientation into account.  Orientation index is part of the build
plane state, and seems to be working correctly.