WIP Normalized newDesign & loadDesign
Any of the apps should be able to use either a new design (if it
makes sense there) or a loaded design.  This means that I need
controllers created with loaded designs, and possibly some of the
"viewer" events with a new design.

I've hacked things together with RenderHistory attached to
the root controller, but it is pretty awful.  In particular, after
any edit action I'm doing a full batch render, though this is
probably the best approach.

STATUS: all apps are working, except:

1. selecting a ball in classic
2. octahedral symmetry in buildplane is partially working, bizarrely