Implemented legacy file save
I have added `EditHistory.java` to the JSweet set, with a little difficulty
around `DomSerializer.java`, overcome by using injection.  I initially
tried to make a stub of the latter, and exclude the real one from the
JSweet transpile, but somehow I always got "duplicate class" even when
I managed to get the sourceSet exclude right... something to do with
the classPath, I guess.

The worker does not serialize a full vZome document, leaving the
camera and lighting to the client.

The store contains both an original camera that drives the rendering
(for load), and a `liveCamera` that stays in sync with the rendering.
The latter gets serialized.  If I try to use just one value in the store,
I get a dependency loop, with visible effects (unable to zoom).