WIP Undo / Redo for buildplane and classic
Undo and redo are working, but the problem is that the
buildplane state gets out-of-sync... notably, the center can be left
in space with no ball.  I need a general solution that captures the
tool state (for buildplane or any other tool) at the moment an
undoable action is sent to the worker, or perhaps at the moment
the new scene comes back.

I've also added command menus to the classic app, including a
working undo/redo.  This is possible because I upgraded @suid,
which now has Menu and MenuItem.

PENDING: the can-undo and can-redo state has to come back from the
worker, to enable and disable the buttons.

PENDING: classic can open files, but the UI is still listening to PCEs from
the old controllers, so does not respond any more.

Finally, I'm experimenting with key event handling.  Mac shortcuts work
for undo/redo, as long as some event has triggered keyboard focus.