Confined SolidJS to just Classic UI
On the original `integrate-solidjs` branch, I attempted to port everything
to SolidJS.  This stalled because the ecosystem there is still weak; in
particular, `@suid` is not up-to-date with `@material-ui`, and there is
no real equivalent of `@react-three/fiber`.

Therefore, here I'm confining the use of SolidJS to just the classic UI,
where I have found it to be most helpful for mapping to the legacy
Controller concepts.  In the classic app, I use `solidify()` in a couple
of places to use React-based components, notably in the appbar and
in the canvas.

One thing I need to correct: `classic/controllers-solid.js` replicates a lot
of the code from `ui/viewer/store.js` in terms of worker interactions,
and it does not implement the entire contract for saving vZome files.

Other than that, this code is very functional, with all apps working
correctly.