Recovered windows-to-dreamhost.sh
This is a script for completing the GitHub Action to deploy a new vZome
desktop build to Dreamhost, which fails to sftp the Windows build
because of some SSL problem on Windows.  I run it on my local machine.

Note that this does not actually publish vZome desktop; I have to edit
the `.htaccess` file to do that, changing the build number for the links.