Some classic improvements
I added the trackball canvas, mostly.  It is sizing itself too large for
some reason, and there is no trackball or camera sync.  To support
this, I added a new `trackball` prop to disable the trackball.

I also added support for `?canDownloadScene=true` query param,
so I can capture ready-to-use scene JSON for symmetry models.
`icosahedral-vef.js` was exported in this way.

STATUS:

1. background color does not sync when a new model is opened

2. Camera canvas is too wide.  I've asked about this on the Poimandres
  Discord server, sharing this (working!) sandbox:

https://codesandbox.io/s/r3f-in-html-layout-c1dbkc