Fixed the online camera save issue
vZome desktop wants to keep a fixed frustum shape, and expects
that when rendering.  vZome Online was not updating the `width`
and `far` camera values to maintain this invariant, resulting in a
broken FOV and fog on the wrong scale.

Worst, however, was the bad distance computation happening
in syncCamera.