WIP Creating classic strut drag tool
I have had a lot of trouble with this.

The events seem to be largely correct, entering and exiting "strutting"
mode correctly.  Finishing the drag back on a ball still selects the ball,
but I can live with that.

Next step is to turn the drag events into rotations.

To get it to work, I had to introduce SceneEditor, since the scene tools
must all be React-based, and setting the `children3d` outside in the
SolidJS code was not working.

I renamed most occurrences of `toolRef` as `toolActions`, to be
more accurate.