WIP Enabling more menu commands in classic
I added the missing setEditorModel() calls for symmetrySystems,
and I removed the useless DefaultController, replacing it with the
EditorController.

I've also added proper error reporting, and added the System menu,
though it is all disabled at the moment.  The Help menu is now there, too.

ReplaceWithShape (tools menu) still does not work, so it is disabled.

I've also updated the Help menu links in desktop, to avoid the old blog.