Completed refactoring of exporters
I finally refactored Exporter3d into GeometryExporter (which only
requires the RenderedModel) and DocumentExporter (which needs
the DocumentModel for any reason).

I also cleaned up the use of SnapshotExporter, using a functional
approach.

The desktop application is working, and the online classic app can
export STL to the console.  I had to implement StringWriter and
PrintWriter for this (or any exporter) to work in Javascript.

I also added a Promise-based exporter function to the SolidJS store.
This is working fine, but could bear some cleanup, or possibly some
generalization.