Add SnubDodecahedronField
Equivalent to the legacy SnubDodecField but derived from ParameterizedField.
Now all of the legacy fields have a comparable ParameterizedField.
This class is not used anywhere yet, but it's been implemented for a long time and I wanted to get it into the code base just for completeness.
This is also a necessary step towards refactoring and merging ParameterizedField into AbstractAlgebraicField.