Classic can now switch symmetry
Global table of PCLs was a problem, so I'm replacing it with
a system of controller peers.  I thought I needed this for some kind
of complicated controller replacement for symmetry changes,
but then I realized I could just switch the subcontroller name in
the UI, and leave the controllers (and their new wrappers) unchanged.
Even so, the system of wrappers is much more understandable.

I ran into an issue with the legacy code, when the symmetry change
triggers re-rendering for new shapes.  There was a Thread.yield(),
since the code was still running on the EDT in Swing.  I added
an ExclusiveAction there (as suggested by a TODO) to move that
code to a worker, removing the necessity for the yield.

Most importantly, the ClassicEditor UI code now uses the "symmetry"
top-level property to select a subcontroller, e.g. "symmetry.icosahedral".
Naturally SolidJS makes this all work smoothly, and I didn't need any
special "re-sync" code in the ControllerWrapper in the worker, since
all the subcontrollers and their properties get re-fetched automatically.