WIP Restoring classic features in SolidJS
Canvases have the right initial size, largely, and resize seems to work,
including correct computation of aspect ratio and fov.

Lighting seems to be working, but it is hard to tell without TrackballControls.
The PerspectiveCamera is working now, after realizing that `makeDefault` was
a Drei idiom only.

The silent failures in reactive code are very frustrating... but if I stop on caught
exceptions I can see them.