Custom Menu macros
This supports custom menu entries of the form:
```
MACRO/SelectAll,icosasymm,pointsymm,DeselectAll=Full Icosa Everything ^2
```
It only supports global, built-in action strings.  It works by simply executing them in
sequence; to support this, ExclusiveAction now uses a SingleThreadExecutor, which
is a good idea anyway.

Each command is individually recorded in the history; the DocumentModel does not
have any knowledge of the macros, since they are simply executed at the UI level.

It is of very limited use, without even access to predefined tools.  It could perhaps
find more application with non-edit actions, such as exports, but even there it is
limited in scope, since it cannot engage the UI at all.