WIP Ported other apps & web component to Solid
The `bhall/basic` and `fivecell` apps seem to be working, though they cannot
load the models on this machine, as usual.

The `DesignViewer` being used is still skeletal, missing all the UI elements like
the spinner and error banner, and the icon buttons.

I implemented a stand-in for SUID `Tabs` component, which has not yet been
ported from MUI.  It uses a Select.

I haven't taken any steps yet for the web component, which will use `solid-element`.