More functional parity in Solid DesignViewer port
Created stub for SUID Tooltip.

Spinner and ErrorAlert are working correctly.  Viewer app works correctly,
including the "open" menu icon.

The "inspector" mode for the viewer is still disabled until I port the
DesignHistoryInspector component.

Browser app and the web component are still not ported from React.

The camera is not getting set correctly when a file is loaded... possibly for a vZome
file only.