Converting web component from MUI to Kobalte
Spinner, ErrorAlert, ScenesMenu, Exports button, and Fullcreen button have been ported.  Still
remaining: XR preview, undo/redo, settings.

The Exports menu is functional, but requires an extra tab key or a click to function.  Export is only supported for vZome files at the moment. It is possible to get into a state where the menu just
flashes, when playing with keyboard focus using the tab key.

Also, I may have broken classic "save" with the controllers-solid.js change.

The Scenes menu could use some CSS tweaking, and has an odd behavior on
first click.

I used the styles from Kobalte examples, with some adjustments of my own,
like making the alert background red, and aligning SVG icons in buttons correctly.

There is a defect in the alert: when one alert is dismissed, all others on the page
get dismissed too.

Spinner is now using a Kobalte Progress.  I need to look into CSS to make it a
circular progress again.

Kobalte Portal elements are always mounted to the UrlViewer root, which works
well for a Solid app like `fivecell` and also for the web components.

The `urlViewerCSS` style element is added to the shadow root for the web
component, but in other cases it is added (once!) to the document body.

All SUID dependency is removed from the core web component, and I did some further
cleanup of exports to rationalize the splitting chunks a bit more.  I understand what esbuild
seems to be doing, now, I think.