Final CSS, camera FOV, and lighting tweaks
Browser and classic are now using the idiom found by Vincent (bigmistqke on
Solid JS Discord) to resize the canvas correctly without getting into a loop
fighting with CSS for the grid: each div between the grid and the canvas
must have overflow=hidden, display=flex, and height=100%.

Web component camera & lighting seem to be correct.  There is still a background
color change when using the scenes menu for vZomeLogo.