Length panel full function
The up/down buttons and the scale preset buttons, the half switch, the scale factor,
the slider, and the length display are all working.

The scale factor is not fully generalized for fields with higher orders.

The slider does not support any direct manipulation.

BIGGEST TIME SINK: forgetting that 'true' is not a boolean value!  Was also the problem
earlier with the "single" switch for the orbit triangle, now fixed.

I had to add support for setProperty() in the controller client.

I also fixed a defect in setSingleOrbit, which was exiting without a PCE.