Scene attribute on web component
This will allow programmatic scene selection.  Scene titles are discoverable by subscribing
to the "vzome-scenes-discovered" CustomEvent, emitted by the viewer element.
See `online/public/test/index.html` for sample code.

A scene index, rather than a title, can be selected as "#6" for index 6, zero-based.

Invalid title or index will result in a console warning message, and display of the default scene.

The `show-scenes` attribute is irrelevant when the scene attribute is set; the viewer is
thus scene-controlled and the scenes menu won't be displayed.