Corrected scene attr tracking for web component
The attr change was not being observed.

To make this work, I switched SELECT_SCENE to take a title not an index.  Obviously,
anonymous scenes must be selected with "#n".

There is still an issue with the scenes menu not disappearing when the scene attr is set.
See vzome-viewer.js, line 104.