Fixing the use of XML entities in scene titles
Since the user can enter any text into the article page titles in desktop vZome, the
serialized XML may contain entities that encode those characters, e.g. "&" becomes "&amp;".
The XML parser normally reverses this, but the `txml` parser I'm using in the worker does
not!  This will be even more important when other languages are used.

The fix is simple, and is done in the main context, where I can leverage the DOM parsing
and decoding of entities.  I do have to do the decoding twice, once for each worker
subscriber, and I have to encode when sending SCENE_SELECTED.