Refactored setSymmetrySystem() in Java
The setSymmetrySystem() method is gone.  It was trying to do two things
at once:
 - initialize controllers to match document model state
 - change document model state AND controller when switching symmetry

Now, the new setSymmetryController() method really only does the latter.
When initially constructing the DocumentController, the model's symmetry
is mapped to a SymmetryPerspective *label*, and that is used in a call
to setSymmetryController().

SymmetryPerspective.getLabel() is a new method to support this.  The only
real use-case at the moment is "heptagon antiprism corrected", which is the
name of a symmetry with a corresponding SymmetryPerspective labeled
just "heptagon antiprism".  The Symmetry with the latter name is deprecated, so the
corresponding label is null.