New / Open / Save improvements
Added appbar feedback when there is a filename, and showing:
1. opened from URL (world icon) or disk (disk icon)
2. unsaved changes (highlighted icon)

Undo/redo is not affecting the "edited" state yet.

I've tried to clearly identify store state managed by the client context: 'fileHandle'
and 'liveCamera' are only
set by the client, and 'designName' and 'problem' are managed by client and worker.

There is some bug reopening a saved radome-3 design.