Fixed interpreter branch history bug
The interpreter was simply not adding branches to the history, since it wasn't built
to faithfully reproduce the history for saving again.

I split EditCursor out of RenderHistory, since they really are different functions.
I still want to merge interpret() and EditCursor into "Interpreter", but that can wait.
The EditCursor already cleans up the interpreter step() function a fair bit.

RenderHistory is no longer able to perform interpret() in order to render a scene
that was not captured yet, as when the debugger is in use.  I'll have to do more
refactoring to make that work, and it needs more thought.

RenderHistory could be optimized (for non-debugger use) to only capture scenes
for snapshots and the targetEdit state.

I renamed LegacyEdit as ParsedEdit.

I added a Beta Version warning dialog.

I disabled the handle recording after open, so people don't corrupt existing files
in the event of a save bug.  I also fixed the file name after saveAs, to match the handle.