Context menu
The JSX and CSS is done, working great.  Only "lookAtOrigin" and "lookAtThis" are
connected.

I have separated the camera trackball as a third mouse mode, so that there
is no conflict on right-click.

I have it working for picking scene objects!

I am also using the "rotate" cursor for the camera trackball view, by virtue of a
DummyTool for the Camera view.

Camera near & far still need to be controlled.