Copy / use copied camera contextual menu items
The basics are working fine.

This work revealed a bug: unless in camera mode, opening a file did not render
the scene!  The same happened after "useCopiedView".  I fixed it by controlling
the PerspectiveCamera lookAt even when no trackball is enabled.

I also refactored the TrackballControls createEffects, since the listener setup was
happening on any lookAt change, and maybe others.