Added OpenSCAD and Python build123d mesh exporters
I also "fixed" the STEP exporter so it does not fail silently.  It still requires the user
to create a file, but then the file is only used to get its parent directory, and then
one file per shape is exported.

The STEP exporter is rather useless as-is, but could be turned into a single-manifold
export for color 3d printing.

I reorganized the "Export 3D..." submenus a bit, for more clarity.

I also started using Resources.loadStringResource() in GeometryExporter
getBoilerplate(), so that exporters will work in Online also.