Adding more exporters to Online
In order to add the OpenSCAD and build123d exporters to Online, I needed to make the
necessary boilerplate resources available in Online.  I added a build-online.bash step to
index the necessary resources into a new field in the generated revision.js.  The indexed
resources are loaded and injected into core as soon as the baseURL is known.

All menu items are enabled except Collada, POV-Ray, and vZome Shapes JSON.