WIP Capture vector graphics for Online
I've refactored so that Java2dSnapshot and Java2dExporter have no dependency
on javax.vecmath.  I have done the same for Lights, simultaneously removing the old
getter with a side-effect.  For Camera, I decoupled by adding methods to fetch the
view and eye transforms as RealMatrix4; the actual transforms are computed in static
methods of RealMatrix4, so they can be reused in Javascript.

On the JS side, I'm passing scene.liveCamera across to all exporters (running in the worker).

I've incorporated a branch that improved the code for the Camera panel tracking the main
scene camera and lighting changes.  I've also FINALLY fixed the near and far clipping planes,
keeping the frustum shape constant at different scales (camera distances).

STATUS: Java refactoring is working fine in desktop.  The shims for java.awt.* are all in place
for the JSweet transpile.  All the exporters transpile successfully.

There are no runtime error in Online, and in general the SVG output is correct.  Until the
trackball has been rotated, apparently the near&far values can be wrong, or perhaps the
distance.  I still don't know why aspectRatio must be 1.0 rather than width/height.