Refactored camera state into CameraContext
RotationProvider is now renamed CameraProvider, and the context offers two new methods.
The `adjustFrustum` method is called on every trackball change (live), to keep the frustum
shape fixed as in desktop.  The `recordCamera` method is used to keep `liveCamera`
in sync with the three.js camera.

I've also fixed `liveCamera` to have the correct state after loading a scene, so that vector
(SVG, etc.) capture now works correctly in all cases.