Adding MathML support
Every Field.Element (like AN or BR) now supports `getMathML`.  The result is ready to
be placed inside `<math>...</math>` or any other bit of MathML.  A `<mrow>` is
automatically added when necessary.

Along the way, I corrected the use of `getIrrational` in the JS field implementations.

Finally, the MathML output is now used in Online classic, to display the final strut length.