Restored glTF exporter in vzome-viewer
I had to update `three-stdlib` to get a version of the exporter that worked.  I tried updating
`three.js` also, but it changed the lighting, so I reverted that.

The glTF export mostly works, but the single `buffer` object comes out with just a `byteLength`
property and nothing else... no data.  I'll have to step through the code to see why.