Cell selection implemented in 59icosahedra
This is a bit of a hack, requiring hand-editing of the .vZome files for the two selector
models to add `label` attributes to the `GroupSelection` commands.  Still, the hack is
limited to the Javascript code, so far.

For the stellation model, I now have named bookmarks for all the stellation faces, and the
first four stellation cells.

I have implemented a very rough "macro capture" to use in Online classic, and a means
to trigger a macro from the client.

The side-effect of clicking on one of the selector cells is still just a log statement, until
I work out how to communicate the macros to the stellation worker.  For the moment,
a background click in the stellation scene will trigger a macro to produce colored "c"
cells.